mvp project version & refactor

+++ src/core/config/http.py
@@ -3,6 +3,6 @@ from pydantic_settings import BaseSettings, SettingsConfigDict
 
 class HTTPSettings(BaseSettings):
     proxy: str = "http://example_user:example_pass@ip:port"
-    timeout: int = 30.0
+    timeout: float = 30.0
 
     model_config = SettingsConfigDict(env_prefix="HTTP_")
